R B and T J
We have use the hashtag in order to thourghly discribe the process
occuring throughout our code. We have successfully created the deck in
the form of a list. We have also shuffled and dealt one card to each
player to decide the dealer and thus who goes first. We have then put
the 3 cards back and dealt five cards to each player. We have begun to
create a while loop based on books and how to determine who goes first.